删除原有tty文件夹，并更改新tty文件名
