--- kernel/src/driver/tty/vt/mod.rs
@@ -1,10 +1,0 @@
-//! Virtual terminal driver.
-//!
-//! (TODO) This driver is not implemented yet.
-
-/// The minimum number of virtual terminals.
-#[allow(dead_code)]
-pub const MIN_NR_CONSOLES: usize = 1;
-/// The maximum number of virtual terminals.
-#[allow(dead_code)]
-pub const MAX_NR_CONSOLES: usize = 63;
