+++ kernel/src/libs/lib_ui/textui.rs
@@ -1,7 +1,7 @@
 use crate::{
     driver::{
         serial::serial8250::send_to_default_serial8250_port,
-        tty_new::{
+        tty::{
             tty_driver::TtyOperation,
             virtual_terminal::{virtual_console::CURRENT_VCNUM, VIRT_CONSOLES},
         },
